refactor: rename score package into scoring
